add: solution for tree, set

+++ acmicpc/1074.cpp
@@ -1,23 +1,23 @@
 #include <cmath>
 #include <iostream>
 using namespace std;
-long long r, c;
-void division(long long N, long long cnt) {
+int r, c;
+void division(int N, int cnt) {
   if (N == 0) {
     cout << cnt;
     return;
   } else {
-    long long div = pow(2, N - 1); // 행 또는 열의 너비
-    long long sum = pow(4, N - 1); // 디비전의 크기
-    if (r < div) {                 // 1,2사분면
-      if (c < div) {               // 2사분면, 누적 x
+    int div = pow(2, N - 1);
+    int sum = pow(4, N - 1);
+    if (r < div) {
+      if (c < div) {
         division(N - 1, cnt);
       } else {
         c -= div;
         division(N - 1, cnt + sum);
       }
-    } else {         // 3,4사분면
-      if (c < div) { // 3사분면
+    } else {
+      if (c < div) {
         r -= div;
         division(N - 1, cnt + sum * 2);
       } else {
@@ -29,7 +29,7 @@ void division(long long N, long long cnt) {
   }
 }
 int main() {
-  long long N;
+  int N;
   cin >> N >> r >> c;
   division(N, 0);
 }
